fix(SHS-6191): refactor based on PR feedback
