+++ docroot/modules/humsci/hs_events/hs_events.services.yml
@@ -1,4 +1,7 @@
 services:
-  hs_events.auto_unpublish:
-    class: Drupal\hs_events\Services\AutoUnpublish
+  hs_events.auto_unpublish_events:
+    class: Drupal\hs_events\Services\AutoUnpublishEvents
+    arguments: ['@entity_type.manager', '@logger.factory']
+  hs_events.event_form_modifier:
+    class: Drupal\hs_events\Services\EventFormModifier
     arguments: ['@entity_type.manager', '@logger.factory', '@config.factory']
