+++ docroot/modules/humsci/hs_events/src/Services/AutoUnpublishEvents.php
@@ -2,15 +2,14 @@
 
 namespace Drupal\hs_events\Services;
 
-use Drupal\Core\Config\ConfigFactoryInterface;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Datetime\DrupalDateTime;
 use Drupal\Core\Logger\LoggerChannelFactoryInterface;
 
 /**
- * Service for auto-unpublishing past events and handling form modifications.
+ * Service for auto-unpublishing past events.
  */
-class AutoUnpublish {
+class AutoUnpublishEvents {
 
   /**
    * The entity type manager.
@@ -25,13 +24,6 @@ class AutoUnpublish {
    * @var \Drupal\Core\Logger\LoggerChannelFactoryInterface
    */
   protected $loggerFactory;
-
-  /**
-   * The config factory.
-   *
-   * @var \Drupal\Core\Config\ConfigFactoryInterface
-   */
-  protected $configFactory;
 
   /**
    * Node type for events.
@@ -49,42 +41,31 @@ class AutoUnpublish {
   const EVENT_DATE_FIELD = 'field_hs_event_date';
 
   /**
-   * Config page ID for site options.
-   */
-  const SITE_OPTIONS_CONFIG_ID = 'hs_site_options';
-
-  /**
-   * Field name for site auto-unpublish setting.
-   */
-  const SITE_AUTO_UNPUBLISH_FIELD = 'field_site_auto_unpublish';
-
-  /**
-   * Constructs a new AutoUnpublish service.
+   * Constructs a new AutoUnpublishEvents service.
    *
    * @param \Drupal\Core\Entity\EntityTypeManagerInterface $entity_type_manager
    *   The entity type manager.
    * @param \Drupal\Core\Logger\LoggerChannelFactoryInterface $logger_factory
    *   The logger factory.
-   * @param \Drupal\Core\Config\ConfigFactoryInterface $config_factory
-   *   The config factory.
    */
   public function __construct(
     EntityTypeManagerInterface $entity_type_manager,
     LoggerChannelFactoryInterface $logger_factory,
-    ConfigFactoryInterface $config_factory,
   ) {
     $this->entityTypeManager = $entity_type_manager;
     $this->loggerFactory = $logger_factory;
-    $this->configFactory = $config_factory;
   }
 
   /**
    * Unpublishes past events that have auto-unpublish enabled.
    *
+   * @param int $limit
+   *   Maximum number of events to process. Defaults to 100.
+   *
    * @return int
    *   The number of events that were unpublished.
    */
-  public function unpublishPastEvents(): int {
+  public function unpublishPastEvents(int $limit = 100): int {
     $node_storage = $this->entityTypeManager->getStorage('node');
 
     // Get current time in UTC to match database storage.
@@ -100,6 +81,7 @@ class AutoUnpublish {
       ->condition('status', 1)
       ->condition(self::AUTO_UNPUBLISH_FIELD, 1)
       ->condition(self::EVENT_DATE_FIELD . '.end_value', $current_timestamp, '<')
+      ->range(0, $limit)
       ->accessCheck(FALSE);
 
     $nids = $query->execute();
@@ -119,7 +101,7 @@ class AutoUnpublish {
         $unpublished_count++;
 
         // Log the action for debugging purposes.
-        $logger->notice('Auto-unpublished event "@title" (ID: @id) as it is past its end date.', [
+        $logger->info('Auto-unpublished event "@title" (ID: @id) as it is past its end date.', [
           '@title' => $node->getTitle(),
           '@id' => $node->id(),
         ]);
@@ -134,70 +116,12 @@ class AutoUnpublish {
     }
 
     if ($unpublished_count > 0) {
-      $logger->info('Auto-unpublished @count past events during cron run.', [
+      $logger->info('Auto-unpublished @count past events.', [
         '@count' => $unpublished_count,
       ]);
     }
 
     return $unpublished_count;
-  }
-
-  /**
-   * Checks if the site auto-unpublish setting is enabled.
-   *
-   * @return bool
-   *   TRUE if the site setting is enabled, FALSE otherwise.
-   */
-  public function isSiteAutoUnpublishEnabled(): bool {
-    try {
-      $site_options = $this->entityTypeManager
-        ->getStorage('config_pages')
-        ->load(self::SITE_OPTIONS_CONFIG_ID);
-
-      if (!$site_options) {
-        return FALSE;
-      }
-
-      $field_value = $site_options->get(self::SITE_AUTO_UNPUBLISH_FIELD);
-      return $field_value && $field_value->value == '1';
-    }
-    catch (\Exception $e) {
-      // Log the error but don't break the form.
-      $this->loggerFactory->get('hs_events')->error('Error checking site auto-unpublish setting: @error', [
-        '@error' => $e->getMessage(),
-      ]);
-      return FALSE;
-    }
-  }
-
-  /**
-   * Sets the default value for the auto-unpublish field based on site setting.
-   *
-   * @param array &$form
-   *   The form array to modify.
-   */
-  public function setAutoUnpublishDefaultValue(array &$form): void {
-    // Check if the auto-unpublish field exists in the form.
-    if (!isset($form[self::AUTO_UNPUBLISH_FIELD]['widget']['value'])) {
-      return;
-    }
-
-    // Set the default value based on site configuration.
-    if ($this->isSiteAutoUnpublishEnabled()) {
-      $form[self::AUTO_UNPUBLISH_FIELD]['widget']['value']['#default_value'] = 1;
-    }
-  }
-
-  /**
-   * Moves the auto-unpublish field to the options group.
-   *
-   * @param array &$form
-   *   The form array to modify.
-   */
-  public function moveAutoUnpublishFieldToOptions(array &$form): void {
-    if (isset($form[self::AUTO_UNPUBLISH_FIELD])) {
-      $form[self::AUTO_UNPUBLISH_FIELD]['#group'] = 'options';
-    }
   }
 
 }
